mcp-server/requirements.txt
fastmcp>=0.2.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
fastapi>=0.68.0
uvicorn>=0.15.0